+++ test/h5cpp_tests/__init__.py
@@ -1,6 +1,6 @@
-from .file_creation import *
-from .file_image import *
-from .file_open import *
+from .file_creation_test import *
+from .file_image_test import *
+from .file_open_test import *
 from .property_tests import *
 from .group_tests import *
 from .datatype_tests import *
@@ -8,5 +8,5 @@ from .attribute_tests import *
 from .node_tests import *
 from .filter_tests import *
 from .dataspace_tests import *
-from .path_tests import *
+from .path_test import *
 # from .group_construction import *
