+++ test/h5cpp_tests/dataspace_tests/__init__.py
@@ -1,2 +1,2 @@
-from .scalar_dataspace import *
-from .simple_dataspace import *
+from .scalar_dataspace_test import *
+from .simple_dataspace_test import *
