+++ test/h5cpp_tests/datatype_tests/__init__.py
@@ -1,1 +1,1 @@
-from .predefined_type_tests import *
+from .predefined_type_test import *
