+++ test/h5cpp_tests/file_image_test.py
