+++ test/h5cpp_tests/filter_tests/creation_test.py
@@ -136,17 +136,17 @@ class FilterCreationTest(unittest.TestCase):
 
     def testScaleOffset(self):
 
-        filter = ScaleOffset(SOScaleType.INT, 2)
-        filter(self.dcpl)
+        sfilter = ScaleOffset(SOScaleType.INT, 2)
+        sfilter(self.dcpl)
         hdf5.node.Dataset(self.root, hdf5.Path("ScaleOffset"),
                           self.datatype,
                           self.dataspace,
                           self.lcpl,
                           self.dcpl)
         self.assertEqual(self.dcpl.nfilters, 1)
-        self.assertTrue(filter.is_encoding_enabled())
-        self.assertTrue(filter.is_decoding_enabled())
-        self.assertEqual(filter.id, 6)
+        self.assertTrue(sfilter.is_encoding_enabled())
+        self.assertTrue(sfilter.is_decoding_enabled())
+        self.assertEqual(sfilter.id, 6)
         filters = ExternalFilters()
         self.assertEqual(len(filters), 0)
         flags = filters.fill(self.dcpl)
@@ -237,7 +237,8 @@ class FilterCreationTest(unittest.TestCase):
             except RuntimeError:
                 error = True
             if not error:
-                print("filter with 32008 id  created")
+                pass
+                # print("filter with 32008 id  created")
             # for newer versions of hdf5 you can constuct the object
             #       on nonexisting filter
             # self.assertTrue(error)
@@ -261,7 +262,8 @@ class FilterCreationTest(unittest.TestCase):
             except RuntimeError:
                 error = True
             if not error:
-                print("filter with 31999 id  created")
+                pass
+                # print("filter with 31999 id  created")
             # for newer versions of hdf5 you can constuct the object
             #       on nonexisting filter
             # self.assertTrue(error)
