+++ test/h5cpp_tests/group_tests/__init__.py
@@ -1,1 +1,1 @@
-from .group_construction import *
+from .group_construction_test import *
