+++ test/h5cpp_tests/node_tests/__init__.py
@@ -1,7 +1,7 @@
 from .dataset_io_test import *
 from .dataset_partial_io_test import *
 from .dataset_direct_chunk_test import *
-from .group_creation_tests import *
+from .group_creation_test import *
 from .dataset_extent_test import *
 from .h5py_compat_test import *
 from .copy_tests import *
