+++ test/h5cpp_tests/property_tests/__init__.py
@@ -1,4 +1,4 @@
-from .object_copy_list import *
-from .copy_flags import *
-from .chunk_cache_parameters import *
-from .creation_order import *
+from .object_copy_list_test import *
+from .copy_flags_test import *
+from .chunk_cache_parameters_test import *
+from .creation_order_test import *
