+++ test/nexus_tests/__init__.py
@@ -1,4 +1,4 @@
-from .file_tests import *
-from .factory_tests import *
-from .xml_tests import *
-from .path_tests import *
+from .file_test import *
+from .factory_test import *
+from .xml_test import *
+from .path_test import *
