+++ test/regression_tests/__init__.py
@@ -1,3 +1,3 @@
-from .issue_23 import *
-from .issue_18 import *
-from .issue_5 import *
+from .issue_23_test import *
+from .issue_18_test import *
+from .issue_5_test import *
